requirements.txt
Django==1.11
GitPython==2.1.3
flake8==3.3.0
django-extensions==1.7.8
django-crispy-forms==1.6.1
django-multiselectfield==0.1.5
django-autoslug==1.9.3
django-jsonfield==1.0.1
django-background-tasks==1.1.10
ptpython==0.39
